perf(elvis): move the ref of layouts and values out of widgets
