+++ crates/shared/src/serde/mod.rs
@@ -1,3 +1,7 @@
+//! Serde node to target platforms
+#[cfg(feature = "web")]
+mod web;
+
 /// data format transfer
 pub trait Serde<S, T, E> {
     /// Deserialize
