+++ crates/shared/src/tree/mod.rs
@@ -1,10 +1,9 @@
+mod life;
 mod node;
 mod state;
 
 pub use {
+    life::LifeCycle,
     node::Node,
     state::{FnBox, State},
 };
-
-#[cfg(feature = "web")]
-mod web;
